tests/wpt/tests/html/syntax/parsing/resources/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): adoption-agency-1.3
(1,17): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<a><div><style></style><address><a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,35): unexpected-start-tag-implies-end-tag
(1,35): adoption-agency-1.3
(1,35): adoption-agency-1.3
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|         <style>
|       <address>
|         <a>
|         <a>

#data
<nobr><table><marquee></table><nobr>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,22): unexpected-start-tag-implies-table-voodoo
(1,30): end-tag-too-early-named
(1,36): unexpected-start-tag-implies-end-tag
(1,36): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <nobr>
|       <marquee>
|       <table>
|     <nobr>
